+++ packages/synapse-interface/slices/portfolio/updater.tsx
@@ -7,6 +7,7 @@ import { TransactionsState } from '../transactions/reducer'
 import { fetchAndStoreSingleNetworkPortfolioBalances } from './hooks'
 import { PendingBridgeTransaction } from '../transactions/actions'
 import { BridgeTransaction } from '../api/generated'
+import { checkTransactionsExist } from '@/utils/checkTransactionsExist'
 
 export default function Updater(): null {
   const dispatch = useAppDispatch()
@@ -15,14 +16,14 @@ export default function Updater(): null {
     pendingBridgeTransactions,
     userHistoricalTransactions,
     isUserHistoricalTransactionsLoading,
+    fallbackQueryHistoricalTransactions,
   }: TransactionsState = useTransactionsState()
 
   // Update Origin balances when transaction resolves
   useEffect(() => {
     if (!address || !pendingBridgeTransactions) return
     if (
-      pendingBridgeTransactions &&
-      pendingBridgeTransactions.length > 0 &&
+      checkTransactionsExist(pendingBridgeTransactions) &&
       pendingBridgeTransactions.every((obj) => obj.isSubmitted === true)
     ) {
       const updateOriginBalancesForNewestTransaction = async () => {
@@ -41,10 +42,10 @@ export default function Updater(): null {
     }
   }, [pendingBridgeTransactions, address, dispatch])
 
-  // Update Destination balances for new historical transaction
+  // Update Destination balances for new historical transactions
   useEffect(() => {
     if (!address || isUserHistoricalTransactionsLoading) return
-    if (userHistoricalTransactions && userHistoricalTransactions.length > 0) {
+    if (checkTransactionsExist(userHistoricalTransactions)) {
       const updateDestinationBalancesForLastTransaction = async () => {
         const lastTransaction: BridgeTransaction = userHistoricalTransactions[0]
         const destinationChainId: number = lastTransaction.toInfo?.chainID
@@ -58,6 +59,30 @@ export default function Updater(): null {
       updateDestinationBalancesForLastTransaction()
     }
   }, [userHistoricalTransactions, address])
+
+  // Update Destination balances for new fallback historical transactions
+  useEffect(() => {
+    if (
+      !address ||
+      !checkTransactionsExist(fallbackQueryHistoricalTransactions)
+    )
+      return
+    if (checkTransactionsExist(fallbackQueryHistoricalTransactions)) {
+      const updateDestinationBalancesForNewestTransaction = async () => {
+        const newestTransaction: BridgeTransaction =
+          fallbackQueryHistoricalTransactions[0]
+        const updateChainId: number = newestTransaction?.fromInfo?.chainID
+
+        dispatch(
+          fetchAndStoreSingleNetworkPortfolioBalances({
+            address: address as Address,
+            chainId: updateChainId,
+          })
+        )
+      }
+      updateDestinationBalancesForNewestTransaction()
+    }
+  }, [fallbackQueryHistoricalTransactions])
 
   return null
 }